requirements.txt
fastapi
uvicorn
openai
psycopg2-binary
